first approach of new version
This is the first approach for the new architecture. It not only introduces breaking changes but a whole new approach on how to compose components. Every component can now have multiple models/templates and the instance itself has its own template. Also "subscribe" and "eventListeners" are used completely different and "listeners" has a new structure (as it's used for multiple models including $this). HTML and js of components are now much simpler and more structured.